requirements.txt
# Resume Analyzer - Dependencies
flask
scikit-learn
nltk
PyPDF2
python-docx
fpdf2
